refactor: RecentVerificationCard
